controller/requirements.txt
aiofiles==23.2.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
base45==0.4.4
black==23.7.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
colorzero==2.0
cryptography==38.0.4
cssselect2==0.7.0
dnspython==2.4.2
docstring-parser==0.15
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.0
fastapi-limiter==0.1.5
fonttools==4.42.0
gpiozero==1.6.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
orjson==3.9.4
packaging==23.1
pathspec==0.11.2
piccolo==1.0a1
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pydyf==0.7.0
pyotp==2.9.0
pyphen==0.14.0
pypng==0.20220715.0
pyserial==3.5
pyserial-asyncio==0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
requests==2.31.0
rsa==4.9
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
targ==0.3.8
tinycss2==1.2.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
weasyprint==59.0
webencodings==0.5.1
websockets==11.0.3
zopfli==0.2.2
